[FEATURE] Keep gitkeep files when emptying a directory
Both `stencil-app-starter` and `stencil-component-starter` launch
concurrently `stencil build --dev --watch` and `stencil-dev-server`.
This was causing a problem where launching it twice was required to see
the built app.

The first approach was to add a `.gitkeep`, and though that requires some
more tweaking (see https://github.com/ionic-team/stencil-app-starter/pull/50), it is a good one. Sadly, compiler completely empties the output targets
thus removing the `.gitkeep` file.

As using this kind of file is common in the git community, this PRs checks
for the existence of such a file and recreates it after emptying the dir.