fix(esm): use source target to build global esm path
es5 was being use regardless of whether the source target was
actually set to es5. This meant that if source target was set
to es5 the global files generated in generateBrowserCoreEsm(...)
and generateBrowserCoreEs5(...) would share the same path. The
content of file would therefore depend on which executed last. So
you may or may not have got an es5 or es2015 verison of the file.