chore(ci): add github actions workflow
This runs everything except the karma test, because I'm not yet sure how to setup browserstack.

Also, all the tests are run within the same job, which means the indiviual tests (i. e. dist, end-to-end, jest, etc) don't run in parallel. This can be split into multiple jobs, however I'm not yet sure how well that works with respect to the matrix setup.