fix(runtime): Do not double render when attaching cloned elements with Shadow DOM turned off
- In the edge case that we don't use Shadow DOM and have an cloned hydrated component that is (re-)added to the DOM,
make sure that all internal markup is removed, because Stencil will see this component as a new component and will re-add its internal markup as slot content.
This causes perceived double rendering (although technically it is correct) and this commit fixes that by keeping state by adding an 's-isc' attribute. An attribute is needed because the cloning strips off any non-native props such as comp['s-sn'].