feat(componentDidUnload): use disconnectedCallback instead of componentDidUnload
When Stencil is used within other frameworks, elements may be reused, making it impossible for componentDidUnload to be accurate 100% of the time if it is disconnected more than once. Instead, disconnectedCallback() is the preferred way to always know if a component was disconnected from the DOM.

Note that the runtime still works for any collections that have been built with componentDidUnload(). However, updates to Stencil 2 will require it's changed to disconnectedCallback().