refactor(dist): update dist entry extensions
To ensure the extensions are built for the future and work with today's bundlers, we've found it best to use `.cjs.js` extension for CommonJS files, and `.js` for ESM files, with the idea that cjs files will no longer be needed some day, and the ESM files are the standard. (We were using `.mjs` files, but not all of today's tooling and bundlers work well with that extension).

If you're using the `dist` output target, update the `package.json` in the root of your project, like this:

```diff
  {
-     "main": "dist/index.js",
+     "main": "dist/index.cjs.js",

-    "module": "dist/index.mjs",
+    "module": "dist/index.js",
  }
```