chore(ionitron): remove 'awaiting reply' on comment from issues (#3046)
this PR is the first step in trying to improve our automation in GitHub.
it removes the 'Awaiting Reply' label and replaces it with the 'Reply
Received' label when an individual replies to a GitHub issue via a
comment.

this PR also prevents issues that are tagged with 'Reply Received' from
being culled by the stale issue configuration.

this is a first step in improving our workflow. this PR does not cover
every possibility/scenario for interacting with the community.