chore(sourcemaps): minification with sourcemaps (#3093)
this commit introduces a new type, OptimizeJsResult, to be used to name
the shape of the object that is returned by various methods that are
used to minify a user's JavaScript. using this type in the return
signature of methods + in the declaration of objects allows us to
eliminate unnessary type assertions on the individual fields of said
objects.

* add JSDoc to minification to the functions that were investigated as a
part of this effort. this is done in an effort to make the dissemination
of code within Stenicl slightly better. of course, we incur the cost of
maintaining these comments.

* account for minification with sourcemaps

sourcemaps in some regard are at odds with attempting to minize and
compress our JavaScript files. Ideally there's no fidelity lost when
compressing JS, but in some cases, we get inconsistent behaviors across
browsers when we have both 'sourcemap' and 'compress' enabled in our
terser configuration.

after diffing through the terser source/docs, I think our best route
forward for now is to simply disable the compress step. I took a sample
Stencil app and did a `diff` of the output where in one scenario we
didn't set compress to undefined, and one where we did set it to
undefined. other than the sourcemaps themselves, the differences were
rather minimal, and I believe this shouldn't have a large impact on the
output