chore(sourcemap): add stencil core sourcemap configuration
this commit sets up the base configuration that will be required to
enable sourcemaps for the stencil compiler. these sourcemaps differ from
the ones that exist today in that they are sourcemaps for the compiler
itself, and are primarily to be used by the Stencil team/anyone hacking
on the compiler.

this commit does _not_ enable sourcemaps for all parts of the
repository. this commit _does_ contain the necessary configurations that
will be shared across multiple commits/PRs to enable sourcemaps on a
directory by directory basis.

included in this commit is:
- updates to CONTRIBUTING.md that explains how to run a debugger in two
  IDEs
- launch configurations for the debugger for VS Code
- enabling sourcemaps in the project's `tsconfig.json` file
- the addition of the rollup-plugin-sourcemaps package. this will be
  used in subsequent commits to take the sourcemaps generated by
  TypeScript and propagate them through the bundling process via rollup