chore(sourcemaps): add sourcemap support for cli/ (#3108)
this commit enables sourcemaps for the cli/ directory of Stencil. 
this commit does not enable sourcemaps for all parts of the 
repository. these sourcemaps differ from the ones that exist 
today in that they are sourcemaps for the compiler itself, and are
primarily to be used by the Stencil team/anyone hacking on the
compiler.

this commit also sets up the base configuration that will be required to
enable sourcemaps across the entire project:
- updates to CONTRIBUTING.md that explains how to run a debugger in two
  IDEs
- launch configurations for the debugger for VS Code
- enabling sourcemaps in the project's `tsconfig.json` file
- the addition of the rollup-plugin-sourcemaps package. this will be
  used in subsequent commits to take the sourcemaps generated by
  TypeScript and propagate them through the bundling process via rollup