chore(volta): add volta to the project
this commit adds [volta](volta.sh) to the project as a means to
declaratively state the versions of npm and Node that are used in the
project.

specifically, this commit adds a `volta` key in `package.json` and
`test/karma/package.json` to declare the versions of both the package
manager and runtime used by both directories.

this commit is justified/up for discussion via the included adr,
'0014-adopt-volta.md'

the contributing guidelines have been updated under the assumption this
commit will be accepted

STENCIL-5: Add Declarative Node Versioning