chore(scripts): always compile scripts/ directory (#3126)
switch the value of the `incremental` flag in `scripts/tsconfig.json` to
false. when using incremental builds, typescript does not check the
output directory to see whether or not a file that would be emitted
exists or not. this can lead to cases where deleting a file may cause it
to be re-emitted when running compile again. this incurs a ~3 
second increase in build time, but will always ensure that the output
is up to date.

refactor scripts/index.js to attempt to recompile scripts in case
require() fails as a last ditch effort to recompile/require the module
should the `--prepare` flag not be passed in during invocation

update the package.json build script to always recompile scripts/ using
the `--prepare` flag. this allows us to remove the tsc.scripts
invocation, since they effectively do the same thing