refactor(karma): relocate www output target karma tests
this commit is a part of a larger effort to provide additional capacity
to stencil's karma tests for running multiple output targets.
specifically, this commit focuses on moving the existing test root from
the output of the www output target to a sub-directory.

Previously, karma tests generated a directory structure like so:
```
karma/
  test-www/  <-- Location of test output
```

This commit creates the following structure when applied:
```
karma/
  test-output/  <-- Each subdirectory contains its output target's results
    test-custom-elements/
    test-dist/
    test-www/
```

this commit also reorders some commands in the karma directory
package.json to aid in readability of scripts

STENCIL-48: Build Out Infrastructure for Custom Element Builds