chore(build): add sourcemap support to compiler/
this commit adds the minimal amount of code necessary to begin to
support sourcemaps in the compiler directory. these sourcemaps are not
used by consumers of the stencil core compiler/package, but are used by
individuals working on the compiler itself. this commit enables setting
breakpoints, inspecting variables, etc. for the compiler directory
alone.

this commit updates (and renames) the 'hackReplace' rollup transform
hook to output sourcemaps for this transformation step. this is
necessary as the hook is actively changing compiler code, which means
that a new intermediate sourcemap must be generated for rollup to be
able to generate a final sourcemap.

it also enables sourcemaps for the rollupCommonJs plugin for the same
reason.