remove moduleDebugPlugin
this commit removes `moduleDebugPlugin`. this plugin (was) responsible
for generating a comment before every bundled module like so:
```
// MODULE: <PATH_TO_MODULE>
<module_code>
...
```

`moduleDebugPlugin` is removed for two reasons:
1. sourcemaps could not be properly generated for them
2. such comments are largely unnecessary with sourcemaps

the latter point makes the assumption that folks will largely be
debugging in a browser, which is a fair critique that these comments are
nice to have in a browser setting.