chore(build): add sourcemap support to compiler/  (#3142)
this commit adds the minimal amount of code necessary to begin to
support sourcemaps in the compiler directory. these sourcemaps are not
used by projects using the stencil core compiler/package, but are used by
individuals working on the compiler itself. this commit enables setting
breakpoints, inspecting variables, etc. for the compiler directory
alone.

this commit updates (and renames) the 'hackReplace' rollup transform
hook to output sourcemaps for this transformation step. this is
necessary as the hook is actively changing compiler code, which means
that a new intermediate sourcemap must be generated for rollup to be
able to generate a final sourcemap. it also enables sourcemaps for the 
rollupCommonJs plugin for the same reason.

this commit removes `moduleDebugPlugin`. this plugin (was) responsible
for generating a comment before every bundled module like so:
```
// MODULE: <PATH_TO_MODULE>
<module_code>
...
```

`moduleDebugPlugin` is removed for two reasons:
1. sourcemaps could not be properly generated for them
2. such comments are largely unnecessary with sourcemaps

the latter point makes the assumption that folks will largely be
debugging in a browser, which is a fair critique that these comments are
nice to have in a browser setting. if this decision proves to be problematic,
this change can be backed out

add documentation to compiler related plugins

STENCIL-165: Implement Developer Sourcemaps - Compiler