Fix Node.contains mock to match itself
The `Node.contains` method should return `true` if the passed node is the parent/context node.  [The wording in the spec](https://dom.spec.whatwg.org/#dom-node-contains) is that it should return `true` if the passed node is an "inclusive descendant", which differs from a regular "descendant" in that it includes the object itself (not just the child nodes).