chore(jsdoc): add jsdoc to parts of logging system
this commit adds JSDoc to a portion of Stencil's logging infrastructure.
it was motivated by a spike wherein investigating adding error codes to
the compiler were explored. this work is recovered from that effort, to
help aid in understanability of this portion of the codebase.

some methods have been renamed to better reflect their intent.

some usages of `var` have been converted to `let`, as usage of the
variable defined using `var` did not appear to need to be declared using
the keyword