update fn signature of mock-doc's customElementRegistry#whenDefined
set the return type of `whenDefined` to be
`Promise<CustomElementConstructor>` over `Promise<void>` to meet new
type checking expectations with TS v4.4. this leads to a small
functional change in that the return value for this function when the
internal registry contains an entry for a given `tagName`, the function
now returns the constructor associated with the tag name, rather than
void.