useUnknownInCatchVariables
enable `useUnknownInCatchVariables` in all TSConfig files, with the
exception of the main `tsconfig.json`. Although flipping that bit was
originally going to be a part of the TS 4.4 upgrade, our decision to
upgrade from 4.3 -> 4.5 directly increased scope a bit. That and the
fact there are ~90 places where we want to rework the code, I feel more
comfortable splitting that off into a separate effort