chore(prettier): ignore additional test output directories (#3181)
this commit updates the prettier ignore configuration to use globs for
the `dist` and `www` directories under the root `test/` directory. this
prevents files in the `test/end-to-end` directory (and others) from
being formatted

remove `test/browser-compile` to ensure src files are properly
formatted. the test output for this directory is covered by the new
globs