feat(test): jest 27 support (#3189)
this commit restores running the `npm test` command as a part of the
root level package.json's `test.end-to-end` script (and readds it to
CI).

this commit provides support for jest 27 in stencil. this is
accomplished by no longer pulling default options from the jest-cli
package, which now encapsulates its contents through an `exports` field
in its package.json file. during the course of this work, it was
discovered the same fields/defaults were used with Jest 24, 25, & 26.
Those options have been pulled out into a helper function to be provided
for Jest v24 thru 27 to ensure a consistent experience. this helper will
be removed in a future version of Stencil.

the test runner for consumers is locked to jest-jasmine2 for backwards
compatibility in this commit.

jest-runner exposes its default export differently in jest 27. the
creation of a stencil test runner has been updated to handle the
differences between jest v27 and v26 (and under).

jest, jest-cli, jest-environment-node were all updated to the latest
patch version to test with the latest and greatest versions of the
libraries.

update @types/jest to v27, set as recommended version 

add an adr for jest 27 work