feat(mock-doc): add simple MockEvent#composedPath() impl (#3204)
this commit adds a simplistic implementation of composedPath.
specifically, it inspects a MockEvent's target, and attempts to
climb the dom to provide the path the event propagates. this
implementation does not take into account closed shadow roots, nor
slots. the reason for which is that unlike a browser, which has far more
knowledge of the dom and event phases, stencil has none of this at the
mock-doc level, making a deep implementation difficult. a fuller
featured version may be considered for a future version of stencil

STENCIL-240: Enhance MockEvent to support composedPath()