chore(ionitron): don't mark feature requests as stale
make issues with the label "Feature: Want this? Upvote it!" exempt from
ionitron marking such issues as stale. it's an acceptable situation
where we get a totally valid issue/feature request, we mark it as such,
but it's not the top of the backlog. we should be performing manual
checks of these for now (although making this smarter in the future to
look at the number of votes/comments and taking those into account would
be ideal)