fix(testing): don't run tests against non-test files
this commit adds test infrastructure for files picked up by jest for
testing purposes. because there is no efficient way to get the
interaction with stencil+jest (or any libary+jest for that matter), we
instead extract the testRegex field out of the stencil `TestConfig` and
perform `RegExp.test()`-based assertions.

remove a single '?' from the `testRegex` field used that was making the
period preceding (test|spec|e2e) optional