chore(tests): allows tests to be written for scripts/
this commit allows tests (files ending in `.spec.ts`) to be written for
files in the `scripts/` directory and placed in a subdirectory. to
accomplish this, the jest configuration has been updated to pick up
tests under `scripts/`. the `tsconfig` for our scripts has been updated
to _not_ compile tests, as that will be handled by the jest
configuration for the project.