chore(release): fix multiple releases added to github release notes
this commit adds fixes a bug where having multiple successive minor or
major releases (in reality, the former) would be appended to the release
notes generated for github. this caused the release utilities to
generate a url so long that certain browsers (e.g. firefox) to not be
able to handle them, causing this aspect of the release to fail