test(prop-parsing): add numeric prop parsing tests
this commit adds tests to `src/runtime/parse-property-value.ts`.
specifically it adds tests to the case where we parse numbers using the
`parsePropertyValue` function.

these tests were originally a part of
https://github.com/ionic-team/stencil/pull/3254, but were removed from
the pull request. the reason for which is the fix implementation in 3254
ultimately went on a different layer of abstraction then
`parse-property-value`. these tests were originally written with a
different fix in mind; rather than scraping them, they've been updated
to fit the current implementation of the function under test

the intent with this pull request is not to cover every corner case of
`parseFloat`, which is the underlying implementation for this section of
the code. rather it is to document how the function under test will
respond to inputs across different types and of different numeric values