refactor(multiple): code cleanup following webpack treeshaking fix (#3249)
This commit adds several small refactorings that were added as a part of 
https://github.com/ionic-team/stencil/pull/3248. We chose a separate PR
here in that we did not feel these changes helped one understand #3248
in a meaningful way (and may have led to more churn/confusion). 

this commit adds jsdoc to parts of the compilation process that were
investigated as a part of STENCIL-325, "dist-custom-elements
re-assignment causes treeshaking to fail with webpack"

remove an unused argument from the `generateEntryPoint` function for the
`dist-custom-elements` output target

remove an unused argument to the helper function for generating virtual
modules for the `dist-custom-elements` output target

remove the `exp` local variable from the helper function for generating
entrypoints for the `dist-custom-elements` output target. this refactor
is believed to be safe as:
- the array is created locally within the helper function
- nothing is ever added/removed from the array
- the return value of the function is the result of spreading the string
  values in the array into a new array. since this array is always
  empty, this becomes a no-op

update existing generation of `proxyCustomElement` to use newer
`ts.factory` based calls. the replaced calls are drop in replacements
for their predecessors

rename the function for retrieving custom transformers for the
`dist-custom-elements` output target, as it was previously referring to
the `dist-custom-elements-bundle` output target