chore(typings): update browser-compile typings
this commit is a quick fix for the browser-compile tests that currently
do not compile after setting `useUnknownInCatchVariables` in
https://github.com/ionic-team/stencil/pull/3211. This went unnoticed as
this application is not currently type checked by our CI process. In
order to get the application to compile, we type the error as `any` for
now, with the intent of properly narrowing it in STENCIL-371 (and adding
proper CI checks there as well). this commit only unblocks next monday's
release.