chore(test): wire test/ to ci
this commit adds building most projects in the `test/` dir to stencil's
ci system. the `test/package.json` is updated to include the
`browser-compile` subdir, as well as a few script renames. those script
names are propagated up to the project's root level `package.json` so
that they may be called one of two ways:

1. in ci via `test.analysis`
2. locally as a part of the 'test all' script, `test.prod`