chore(karma): remove deprecation warnings in karma tests (#3262)
When we upgraded Stencil to internally use Jest 27
(https://github.com/ionic-team/stencil/pull/3171), we began to see warnings
in the console that look something like:

> ERROR: 'DEPRECATION: An asynchronous before/it/after function took a 
done callback but also returned a promise. This is not supported and will
stop working in the future. Either remove the done callback (recommended)
or change the function to not return a promise. (in spec: SPEC NAME)'

This commit removes either the usage of the `done` parameter or `async`
 keyword from tests, depending on what was required/based on the 
structure of the test itself