chore(karma): upgrade karma to v6 (#3264)
this commit upgrades stencil's version of karma to v6. this is being
done for security purposes (although security vulnerabilities do not
affect end users) and with the hope that there may be some stability
improvements using v5 over v6.

it's required that `@types/node` v14+ be installed to pass type
checking. at this time, our karma tests run under node v12. instead of
trying to circumvent the type checker, we accept the version mismatch,
as we intend to bump stencil to use v14 within the next few months