refactor(jsdoc): add jsdoc to type generation code
this commit adds jsdoc, clarifying comments, and minor type updates to
the codebase. this work was completed while working through adding
`generateTypeDeclarations` to the `dist-custom-elements` output target,
but was split into a separate commit as to not detract from the work
being completed.

STENCIL-329: Generate types for dist-custom-elements output target