chore(types): update typings for analysis task
this commit updates the typings for the analysis task. the team
re-enabled these tasks/tests (for the most part) in
https://github.com/ionic-team/stencil/commit/6f99212e4e33760793a6f8161265cc1c9350e805.
in main. at the same time, a commit that updates typings for onInput
events was sitting in the v3.0.0-dev branch,
https://github.com/ionic-team/stencil/commit/8316a62f5767639e3b70661a0d6b902c9e6daf0b.

upon merging main into v3.0.0-dev, the build broke. this commit fixes
the typings to align with the latter commit.