chore(types): update typings for analysis task (#3298)
this commit updates the typings for the analysis task. the team
re-enabled these tasks/tests (for the most part) in https://github.com/ionic-team/stencil/pull/3263.
at the same time, a commit that updates typings for onInput
events was sitting in the v3.0.0-dev branch,
https://github.com/ionic-team/stencil/pull/3135

upon merging main into v3.0.0-dev, the build broke. this commit fixes
the typings to align with the latter commit.