chore(github): clean up labels/automation (#3309)
this commit removes labels that no longer have any issues associated
with them in the stencil repo & have been deleted. this was a part of a
larger effort to reduce the overall number of labels in the repository
(see https://github.com/ionic-team/stencil/issues/3297)

add ionitron action for needing reproductions. when a team member adds
the "ionitron: needs reproduction" label to an issue, the issue shall
receive a comment stating that more information is needed to reproduce
the issue

this commit adds instructions for ionitron to close issues that have
seen no activity for 14 days AND have been labeled with 'ionitron: needs
reproduction'. this should prevent 'stale' issues