test(methods): add unit tests to method type generation
this commit adds unit testing to the generation of `@Method` decorated
class members - specifically how we generate type information that will
be used in a type declaration file (`d.ts` file).

this commit changes the interface of the internal `generateMethodTypes`
function, requiring that `ComponentCompilerMeta` be passed to the
function rather than just the `ComponentCompilerMethod` metadata found
on the former object. this change is for a future effort where we will
be using additional information from the component metadata to avoid
naming collisions in type declaration file generation. although the
author would have liked to have made this interface change then, adding
the change now simplifies the future pr.