fix(cli): don't generate files if they would overwrite existing code
When `stencil generate` is run it will now check for the presence of all
files it's going to generate before it does so and, if they are already
present, exit with an error message before overwriting the user's code.

This commit also introduces a spec file for the code generation taskfile
(`src/cli/task-generate.ts`) which tests the basic functionality as well
as the change in behavior introduced in this commit.

STENCIL-401: stencil generate overwrites without warning