test(prop): add unit tests to prop type generation  (#3315)
this commit adds unit testing to the generation of `@Prop` decorated
class members - specifically how we generate type information that will
be used in a type declaration file (`.d.ts` file).

this commit adds configuration for test stub files (ending in .stub.ts
or .stub.tsx) to be excluded from our testing configuration, code
coverage, and final output of stencil

this is to support the several stub creation functions that are used in the
aforementioned tests. they are intentionally be committed in the
directory of the files that use them currently. if this pattern is more
widely adopted by the codebase/team, we will move these stubs out to a
more common location