chore(github): split up github actions
this commit reworks the 'main' github workflow to address the two
points above. it uses the concept of composite actions to allow us to
parallelize and categorize our tests in new ways.

previously, we would run the build, unit tests, e2e tests, etc all in one
job, then run one job for each OS/Node version combination

now, we run a job for the build and format separately. upon completion of
the build job, we run the analysis tests, unit tests, and e2e tests for each
OS/Node version in a separate job. this results in more simultaneous runners
(which we appear to be allowed under the GH Free plan for OSS) and a faster
CI run.

this change no longer builds stencil for each OS/Node version, as doing so is
not representative of how stencil is built today. this saves us additional time
in the pipeline and brings it closer to reality.

this change removes the unused `sys.node` script, which simply echoed 'sys.node'
to the terminal.

this change also removes `npm pack`ing stencil and pushing the tarball to github
actions as a separate step, as this was not used.

this change also fixes a small bug in the e2e tests, where dependencies were
not guaranteed to be installed before running the tests