refactor(types): add clarity to type declaration file generation (#3318)
this commit adds jsdoc to various internal/private stencil types for
finding typescript types and avoiding naming collisions. the purpose of
this commit is to keep refactoring/chores of this nature separate from
core logic changes - this allows this commit to be made separate from
existing pull requests

this commit refactors the `update-import-refs` helper module by adding
jsdoc and renaming existing variables in the name of improving clarity.
this module utilizes multiple data structures that cover similar aspects
of the type resolution process, but can be difficult to keep a mental
model of. the intent of this refactor is to clarify the intent/purpose
of this code.