chore(compiler): remove backwards compatibility with assetsDir option
The `assetsDir` option for the `@Component` decorator was deprecated
some time ago in favor of the `assetsDirs` option, but we have retained
backwards-compatibility with the old option nonetheless. This commit
removes that backwards compatibility.

STENCIL-410: Remove Backwards Compatibility for assetDir Field on @Component

BREAKING CHANGE: setting `assetsDir` on a component will no longer work.
Users should migrate to `assetsDirs` instead.